[jforeign] Basic IRMapper
Introducing IRMapper class which is used to map a set of class files to
the internal representation of skill (a type context). Starting with an
input list (user specifies which classes are to be mapped) the mapper
determines the internal representation of those types and of the closure
over the supertype relation.

The code is not intensively tested yet.