[jforeign] rewrite java -> skill type mapping
Rewrite the IRMapper in order to perform two step mapping: 1) shallow
mapping (or collecting) of types, 2) deep mapping which recreates the
type hierarchy discovered during the shallow step and  (in a future
patch) maps fields.

This seems to be a more sane approach in order to get a clean discovery
process and in order to set the basetype/supertype fields to their
right values.

TODO: the skill name is still wrong