[jforeign] Parse explicit mapping rules
This is the first patch of a series to switch to parser combinators in
the mapping file parser implementation. The requirements for the mapping
file make the parser slightly more complicated so the previous design is
not appropriate.
This patch introduces a new parser based on parser combinators and is
able to parse explicit mapping rules.