[jforeign] Get list of required java type names
In order to analyze all foreign java types that are mapped, it's now
possible to obtain a list of all java type names that are mentioned in
the mapping file.

The JavaForeign runner is slightly adapted to print all the gathered
data to the user (developer).

Converted the ListBuffer of class paths to a List because it is
immutable after the command line args parsing is complete.