[jforeign] GeneralOutputMaker makes field access
Make the java-foreign GeneralOutputMaker aware of a "reflectionMap"
which maps IR types to javassist class representations.

Introduce the functions getterOrFieldAccess and setterOrFieldAccess
which check if there is a getter or setter respectively and return the
code patch that uses them. If there is none, it returns a code patch
that accesses the field directly.